Quick fixes - remove `WebHookMultipleEventMapperConstraint` and `WebHookSingleEventMapperConstraint` - use arrays (not lists) when only `IReadOnlyList<T>` is needed - don't use `GetType()` when creating a `private` `ILogger` implementation
