+++ src/Microsoft.AspNetCore.WebHooks.Receivers/Filters/WebHookReceiverExistsFilter.cs
@@ -38,10 +38,10 @@ namespace Microsoft.AspNetCore.WebHooks.Filters
             _logger = loggerFactory.CreateLogger<WebHookReceiverExistsFilter>();
 
             // No need to keep track of IWebHookSecurityMetadata instances that do not request code verification.
-            var codeVerifierMetadata = metadata
+            _codeVerifierMetadata = metadata
                 .OfType<IWebHookSecurityMetadata>()
-                .Where(item => item.VerifyCodeParameter);
-            _codeVerifierMetadata = new List<IWebHookSecurityMetadata>(codeVerifierMetadata);
+                .Where(item => item.VerifyCodeParameter)
+                .ToArray();
         }
 
         /// <summary>
